requirements.txt
aniso8601==9.0.1
boto3==1.26.57
botocore==1.29.57
cftime==1.6.2
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
netCDF4==1.6.2
numpy==1.24.1
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
